Function tamamlandı. HashMap -> ArrayList ile yer değiştirildi. Test methodları eklendi.
